+++ api/jest.config.js
@@ -4,6 +4,7 @@ const config = {
   testEnvironment: "node",
   testPathIgnorePatterns: ["/node_modules/", "/__mocks__/", "/helpers/", "/fixtures/"],
   testSequencer: require.resolve("./jest-sequencer-alphabetical.js"),
+  preset: "ts-jest",
 };
 
 module.exports = config;