+++ api/src/__tests__/young.test.ts
@@ -1,4 +1,5 @@
-const fetch = require("node-fetch");
+import fetch from "node-fetch";
+
 const request = require("supertest");
 const getAppHelper = require("./helpers/app");
 const { dbConnect, dbClose } = require("./helpers/db");
@@ -23,7 +24,6 @@ const jwt = require("jsonwebtoken");
 const { default: faker } = require("@faker-js/faker");
 const { createReferentHelper } = require("./helpers/referent");
 const getNewReferentFixture = require("./fixtures/referent");
-
 jest.mock("../sendinblue", () => ({
   ...jest.requireActual("../sendinblue"),
   sendEmail: () => Promise.resolve(),
@@ -110,7 +110,7 @@ describe("Young", () => {
           } else if (key === "phase2ApplicationStatus") {
             expect(updatedYoung[key]).toEqual(Array.from(young[key]));
           } else if (["birthdateAt", "createdAt"].includes(key)) {
-            expect(Date(updatedYoung[key])).toEqual(Date(young[key]));
+            expect(new Date(updatedYoung[key])).toEqual(new Date(young[key]));
           } else {
             expect(updatedYoung[key]).toEqual(young[key]);
           }