+++ app/src/scenes/inscription2023/steps/correction/stepEligibilite.jsx
@@ -3,7 +3,6 @@ import { toastr } from "react-redux-toastr";
 import { Redirect, useHistory, useParams } from "react-router-dom";
 import validator from "validator";
 import IconFrance from "../../../../assets/IconFrance";
-import CheckBox from "../../../../components/dsfr/forms/checkbox";
 import Toggle from "../../../../components/dsfr/forms/toggle";
 import plausibleEvent from "../../../../services/plausible";
 import { getCorrectionByStep } from "../../../../utils/navigation";
@@ -22,9 +21,9 @@ import DatePicker from "../../../../components/dsfr/forms/DatePicker";
 import ModalSejourCorrection from "../../components/ModalSejourCorrection";
 import { supportURL } from "@/config";
 import DSFRContainer from "@/components/dsfr/layout/DSFRContainer";
-import SignupButtonContainer from "@/components/dsfr/ui/buttons/SignupButtonContainer";
 import Loader from "@/components/Loader";
 import useAuth from "@/services/useAuth";
+import { SignupButtons, Checkbox } from "@snu/ds/dsfr";
 
 export default function StepEligibilite() {
   const { young, isCLE } = useAuth();
@@ -218,16 +217,24 @@ export default function StepEligibilite() {
     <>
       <DSFRContainer title="Vérifiez votre éligibilité au SNU" supportLink={supportURL + "/base-de-connaissance/phase-0-les-inscriptions"}>
         <div className="space-y-5">
-          <div className="flex-start my-4 flex flex-col">
-            <div className="flex items-center">
-              <CheckBox disabled={!isCLE} checked={data.frenchNationality === "true"} onChange={(e) => setData({ ...data, frenchNationality: e ? "true" : "false" })} />
-              <div className="flex items-center backdrop-opacity-100">
-                <span className={`ml-2 mr-2 ${!isCLE && "text-[#929292]"}`}>Je suis de nationalité française</span>
-                <IconFrance />
-              </div>
-            </div>
-            {error.frenchNationality ? <span className="text-sm text-red-500">{error.frenchNationality}</span> : null}
-          </div>
+          <Checkbox
+            state={error.frenchNationality && "error"}
+            stateRelatedMessage={error.frenchNationality}
+            options={[
+              {
+                label: (
+                  <span className="flex items-center">
+                    <span className="mr-2">Je suis de nationalité française</span> <IconFrance />
+                  </span>
+                ),
+                nativeInputProps: {
+                  disabled: !isCLE,
+                  checked: data.frenchNationality === "true",
+                  onChange: (e) => setData({ ...data, frenchNationality: e.target.checked ? "true" : "false" }),
+                },
+              },
+            ]}
+          />
           <Select
             label="Niveau de scolarité"
             value={data.scolarity}
@@ -270,7 +277,7 @@ export default function StepEligibilite() {
               ) : null}
             </>
           )}
-          <SignupButtonContainer labelNext="Corriger" onClickNext={onSubmit} loading={loading} />
+          <SignupButtons labelNext="Corriger" onClickNext={onSubmit} loading={loading} />
         </div>
       </DSFRContainer>
       <ModalSejourCorrection data={modal?.data} isOpen={modal.isOpen} onCancel={() => setModal({ isOpen: false })} onValidation={modal?.onValidation} />