+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# [1.401.0](https://github.com/betagouv/service-national-universel/compare/v1.400.0...v1.401.0) (2024-04-10)
+
+
+### Bug Fixes
+
+* **app:** 2138 - Fixed error on specialSituations Checkboxes. ([05b6397](https://github.com/betagouv/service-national-universel/commit/05b63977bd8875c700706c22ab71edf0d83d7a1d))
+* **app:** Fixed error on specialSituations Checkboxes - Part 2 ([532238c](https://github.com/betagouv/service-national-universel/commit/532238cc753a4b28fbc721645bf4964a07258ce7))
+
+
+### Features
+
+* **api:** 381 add typescript ([#3715](https://github.com/betagouv/service-national-universel/issues/3715)) ([bd3eda2](https://github.com/betagouv/service-national-universel/commit/bd3eda20ad9a6cacb8078c102d39c737bc518184))
+* **api,app,admin:** 2270 - Add RELEASE environment variable ([#3729](https://github.com/betagouv/service-national-universel/issues/3729)) ([3992ddd](https://github.com/betagouv/service-national-universel/commit/3992ddd95e257fd4ab78c882e19610eeb5c5eb5f))
+* **app, lib:** DSFR adjustments. ([#3698](https://github.com/betagouv/service-national-universel/issues/3698)) ([ba16e5d](https://github.com/betagouv/service-national-universel/commit/ba16e5d30757c541d270ccebe73d6da86d703f76))
+* **misc:** 2256 - Add notion task info to release changelog ([#3724](https://github.com/betagouv/service-national-universel/issues/3724)) ([6ec2677](https://github.com/betagouv/service-national-universel/commit/6ec2677c6fcb22d6a21f85975cfef4d90dd9aba7))
+
 # [1.400.0](https://github.com/betagouv/service-national-universel/compare/v1.399.1...v1.400.0) (2024-04-09)